fix(PDM): extend working granuality by minutes
This was solved by extending two tables (`timesheet`, `work_day`)
with the `worked_minutes` integer column.

Postgres also has e.g. 8-byte `time` datatype, but it's not intended
to store time span, but rather time of a day.

So dedicated 4-byte integers for hours and minutes seems to be
a best solution.