Api skeleton, tests and server startup improvements (#44)
* feat: skeleton for API handlers

* feat: started writing the test skeleton

* feat: tests

* feat: more skeletons for now

* feat: more skeletons && some event_task fixes

* feat: primary company test and template adjustments

* feat: get event tests and necessary templates

* event creationm tests

* feat: main event tests done

* feat: tests for getting tasks

* feat: task creation tests

* feat: task update tests

* feat: event comment creation tests

* feat: get employments per user tests

* feat: get specific employment tests

* feat: employment creation tests

* fix: rewrote create tests to include update and delete for convenience

* feat: reworked main to use a configuration function instead. This will make testing simpler, too.

* fix: adjusted tests to use the new initialization procedure

* fix: merged create, update, delete for event comment

* feat: task comment tests

* feat: get event staff tests

* feat: create_update_delete event staff test

* fix: removed redundant tests

* feat: get all event comments tests

* feat: get all task comments tests

* feat: get subordinate tests

* feat: invalid event staff creation

* feat: get_all_assigned_staff tests

* assigned_staff create_update_delete tests

* fix: create_update_dlete assigned staff

* fix: create_update_delete employment, changed user ID to an unemployed user

* fix: cud assigned_staff

* feat: tests finished aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa

* fix: derive in event template

* fix: formatting

* fix: errors

* fix: more misc. fixes

* fix: formatting :|

* fix: duplicate field

* Add SQL scripts and refine existing ones. (#45)

* Fix(DB): correct start_date_lte_end_date constraints

A wrong operator was used.

* feat(DB): add data population script

This is an initial version, it will be extended a bit more
so that each table contains roughly 2 to 5 rows.

* feat(DB): add script for emptying tables

The script is not perfect, it has an issue with breaking
some constraints.

A better version will be submitted soon.

* feat: added an additional employment

* feat: added a comment for an event

* feat: new users

* feat: eventStaff

* refactor(DB): transform tabs to spaces

* fix(DB): capitalize `employment_contract` values

This prevents script to be executed.

* refactor(DB): finish DB emptying script

Now, it should not trigger any errors anymore.

* feat(DB): add an SQL script for deleting all tables & co

It may be useful e.g. for testing.

* refactor(DB): refine two user records

* refactor(DB): refine contract of two users

* refactor(DB): change employment company of two users

* refactor(DB): put task's `description` to dedicated line

* fix(DB): allow `decided_by` columns to be NULL

This conforms to our ER diagram.

This makes sense as both affected tables have also the
`acceptance_status` which suggests `decided_by` assignment
may be delayed.

* feat(DB): fill DB with additional rows

* refactor(DB): delimit table types by a special line

* refactor(DB): reformat `user_record` insert queries.

To keep it <= 80 chars per line.

* feat(DB): add constraint checking decided_by and status

These constraints make sure that when `decided_by` is NULL,
then `acceptance_status` must be equal to `pending` only
as transition into `accepted` or `rejected` must be decided
by someone and stored in `decided_by` column.

* refactor(DB): specify timestamps mostly w/o microsecond precision

For test data, this makes it more easier.

* refactor(DB): assign timesheet to client without `HPP` contract

We agreed that people with `HPP` does not create any timesheets.

* feat(DB): add a constraint into `timesheet`.

The constraint makes sure a user cannot modify his timesheet
once the `approval_status` goes into `pending` or `accepted`
state.

* fix(PDM, DB): correct 2 FK constraints of `timesheet`

The `timesheet` is supposed to reference `employment` table,
not `user_record` and `company`.

This fix lead to a wrong FK being used in data population script
(`data.sql`). This gets fixed in a next commit.

* fix(DB): set a correct `company` FK for `event_staff` record

The `user3_id` has no employment in `company1_id`, but `company2_id`.

This was revealed only after fixing FKs of `timesheet` (the previous commit).

* refactor(DB): change table order in SQL scripts

Needed to be done in order to not trigger any constraint errors.

* docs(DB): add README with psql instructions

* refactor(DB): move SQL scripts into dedicated dirs

This is needed as SQLX does not seem to support
executing a single regular file, it needs to be given a directory.

Note: Migration code at rust side needs to be fixed.

---------

Co-authored-by: Matej Vavrek <matvav13@gmail.com>

* fix: initial test fixes

* fix: tests

* fix: code block formatting in migrations/readme.md

---------

Co-authored-by: svlc <svlc@inventati.org>