feat(DB): allow `employment.end_date` to be NULL
We agreed, that when someone's employment contract
is not HPP, we might want to have NULL there.

There was an idea to put some `DATE TYPE MAX`, but
I just checked that `postgresql` has no constants for this
so it would be difficult to check. This is why I implemented
it by allowing a NULL value.